// frontend/src/api/_generated/openapi.types.ts
export interface components {
  schemas: {
// ...
    AnalysisRunResults: {
      /** Ates Kpi Results */
      ates_kpi_results?: components["schemas"]["AnalysisResultItem"][] | null;
      /** Derived Quantities */
      derived_quantities?: components["schemas"]["AnalysisResultItem"][] | null;
      /**
       * Status
       * @enum {string}
       */
      status: "ok" | "error";
      /** Message */
      message?: string | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}